requirements.txt
moondream==0.0.6
numpy==2.2.0
opencv-python==4.10.0.84
Pillow
psutil==5.9.8
transformers
torch
huggingface_hub
tqdm
accelerate
einops
streamlit
python-opencv
opencv-python-headless
argparse
